Ensure mock has event emitter methods
On iOS, calling `addListener()` or `removeAllListeners()` on a
`NativeEventEmitter` instance also calls `addListener()` or
`removeListeners()` methods on the native module object passed into it.
If those methods aren't present, `NativeEventEmitter` will throw an
error. This was the case when the native module wasn't present on iOS,
and the mock was used, because the mock didn't contain `addListener()`
or `removeListeners()`.

The fix is to add those methods to the mock when `isEventEmitter` is
true.